pyproject.toml
[tool.poetry]
name = "grapp"
version = "0.1.0"
description = "Create graphs on the go"
authors = ["Mexson Fernandes <mexson.fernandes@preflet.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
numpy = "^1.20.1"
pandas = "^1.2.3"
matplotlib = "^3.3.4"
pre-commit = "^2.11.1"
motor = "^2.3.1"
dash = "^1.19.0"
squarify = "^0.4.3"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
aiofiles = "^0.6.0"
dask = "^2021.4.0"
HermesCache = "^0.8.0"
python-dotenv = "^0.17.1"
mysql-connector = "^2.2.9"
jsonschema = "^3.2.0"
python-bsonjs = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
